requirements.txt
numpy
matplotlib
pandas
openpyxl
scikit-learn
tensorflow-macos
tensorflow-metal
torch
torch-geometric
xgboost
